requirements.txt
#--------- libraries needed to run this code ---------#
nltk==3.5
glob2==0.7
numpy==1.23.4
pandas==1.5.1
requests==2.24.0
Unidecode==1.3.0
tensorflow==2.7.0
scikit-learn==0.23.2
googletrans==3.1.0a0
contractions==0.0.52
beautifulsoup4==4.9.3
python-dotenv==0.15.0